+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.21.6...HEAD) - TBD
+## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.21.7...HEAD) - TBD
+
+## [4.21.7](https://github.com/schemathesis/schemathesis/compare/v4.21.6...v4.21.7) - 2026-06-15
 
 ### :bug: Fixed